[Feature] OS - 17: Removed Line chart experiment module
- Removed line chart module
- Added line chart activity in chart container module
