+++ experiments/chartcontainer/src/main/AndroidManifest.xml
@@ -20,6 +20,11 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".LineChartActivity"
+            android:exported="false"
+            android:label="@string/title_activity_line_chart"
+            android:theme="@style/Theme.YGraphs" />
     </application>
 
 </manifest>
