+++ experiments/chartcontainer/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">ChartContainer</string>
+    <string name="title_activity_line_chart">LineChartActivity</string>
 </resources>
